增加进度条value阻尼属性，防止ui界面卡顿

+++ LeeTeke.WpfControl/Models/IMessageWin.cs
@@ -22,7 +22,7 @@ namespace LeeTeke.WpfControl.Models
         object Content { get; set; }
 
     
-        double ProcessValue { get; set; }
+        int ProcessValue { get; set; }
 
         MessageStatus Status { get; set; }
